pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-research-assistant"
version = "0.1.0"
description = "An MCP server that fetches/summarizes Arxiv papers, manages notes, and publishes to GitHub."
authors = [{ name = "zenitsu0509" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "requests>=2.32",
  "httpx>=0.27",
  "pygithub>=2.3",
  "python-dotenv>=1.0",
  "tenacity>=9.0",
  "jinja2>=3.1",
  "arxiv>=2.1",
  "openai>=1.40; python_version>='3.9'",
  "groq>=0.9",
]

[project.optional-dependencies]
llm = [
  "langchain>=0.2",
  "langchain-openai>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
mcp-research-assistant = "mcp_research_assistant.app:main"
